requirements.txt
beautifulsoup4==4.12.2
blinker==1.7.0
bs4==0.0.1
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==3.0.0
gunicorn==21.2.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.2
python-decouple==3.8
requests==2.31.0
soupsieve==2.5
telegram==0.0.1
urllib3==2.1.0
waitress==2.1.2
Werkzeug==3.0.1
